requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
python-dotenv==1.0.0
lxml==4.9.3
selenium==4.18.1
webdriver-manager==4.0.1
PyPDF2==3.0.1
openpyxl==3.1.2
tqdm==4.66.2
